New interactions, including ability to make someone their house head, remove hooks that a character has on the player, fix for starting a regency, and ability to take an artifact.
Update to the plague doctor custom modifier, to hopefully make it more powerful.
Added a new Full Stomachs modifier, which reduces provisions consumption, and increases provisions gain.

Fixed an issue that was preventing the court artificer character spawner option from working.
Fixed an issue that was preventing the akolouthos character spawner option from working.

Edited the pinned character menu so that list view is the default option, so you can see all the pinned characters, then you can check the checkbox if you want a more detailed view.

Added a loading indicator on the title manager when the province tab is loading the provinces, so there is an indication that it's loading, rather than frozen.